requirements.txt
# Web Framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# HTTP Clients & Async
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.8.0
websockets>=10.0

# Environment & Configuration
python-dotenv>=1.0.0
toml>=0.10.2

# Trading Platforms
MetaTrader5>=5.0.0
ccxt>=4.0.0

# Data Processing & Analysis
pandas>=1.5.0
numpy>=1.21.0
pytz>=2023.3

# Technical Analysis
ta>=0.10.2
yfinance>=0.2.32

# Database
sqlalchemy>=2.0.0

# Messaging & Notifications
python-telegram-bot>=20.0

# Task Scheduling
apscheduler>=3.10.0

# Template Engine
jinja2>=3.1.0

# WebSocket Support
wsproto>=1.2.0

# AI & OpenAI
openai>=1.0.0

# Data Export & Documents
reportlab>=4.0.0
python-docx>=1.1.0

# Visualization
matplotlib>=3.7.0

# Configuration Files
PyYAML>=6.0

# Testing
pytest>=7.4.0

# Web Automation (optional)
selenium>=4.15.0

# Task Scheduling (for data retention)
schedule>=1.2.0

# MQTT Support (for unified tick pipeline)
asyncio-mqtt>=0.11.0

# System & Process Utilities
psutil>=5.9.0

# Web Scraping (for breaking news)
beautifulsoup4>=4.12.0
lxml>=4.9.0